J'ai pas d'idée

+++ src/strategie/ActionRemplissage.java
@@ -2,7 +2,7 @@ package strategie;
 
 public class ActionRemplissage extends Action {
 
-	public ActionRemplissage(int cout) {
+	public ActionRemplissage(double cout) {
 		super(cout);
 		this.typeAction = TypeAction.ACTIONREMPLISSAGE;
 	}
